+++ packages/typescript-estree/src/ts-estree/ts-estree.ts
@@ -392,7 +392,8 @@ export type LeftHandSideExpression =
   | PrimaryExpression
   | TaggedTemplateExpression
   | TSNonNullExpression
-  | TSAsExpression;
+  | TSAsExpression
+  | ArrowFunctionExpression;
 export type Literal =
   | BooleanLiteral
   | NumberLiteral